Support fractional display scaling by fixing its one real bug
The Qt frontend truncated devicePixelRatio with int(), so a 125% display got
scale 1 and the in-viewport text was drawn at its unscaled size. Passing the
ratio through instead crashed the packaged app on the first paint, which is why
_gScaleFactor had been left an int.

Installed the Windows SDK debuggers and got the actual stack rather than
guessing again:

  Cmd_Draw -> ExecutiveDrawNow -> OrthoDoDraw -> CScene::draw
    -> SceneDrawImageOverlay -> SceneOverlayOversizeBorder   (read past `data`)

That function resamples with an integer step, upscale = DIP2PIXEL(1). At 1.25
that rounds back to 1 while the destination is still 1.25x wider than the
source, so the source pointer advances once per destination pixel and walks off
the end of the image. Only the whole-number case is sound, so gate the upscale
path on it and fall through to the plain bordered draw otherwise.

_gScaleFactor becomes a float and display_scale_factor an float setting to
match; commanding.py's viewport call rounds explicitly.

Verified on the packaged exe: 3/3 clean at 1.25 (was 3/3 access violations),
and the ray-image overlay still composites correctly. Integer scales are
unaffected -- at 2 the guard passes and upscaling still happens.

+++ layer5/main.cpp
@@ -53,7 +53,7 @@
 #include "Executive.h"
 #include "Feedback.h"
 
-int _gScaleFactor = 1;
+float _gScaleFactor = 1.0F;
 
 #ifdef _PYMOL_NO_MAIN
